fix gpt access scope import cycle
